restructure import hierarchy
