+++ cs2tracker/app/__init__.py
@@ -1,0 +1,1 @@
+from cs2tracker.app.app import Application  # noqa: F401 # pylint:disable=unused-import
