+++ cs2tracker/scraper/__init__.py
@@ -1,0 +1,3 @@
+from cs2tracker.scraper.scraper import (  # noqa: F401 # pylint:disable=unused-import
+    Scraper,
+)
